refactor/test: extract test utils into crate
It will be shared between integration tests and benchmarks.
