+++ dev_utils/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+authors = ["MaidSafe Developers <dev@maidsafe.net>"]
+description = "PARSEC development utilities"
+documentation = "https://docs.rs/parsec"
+homepage = "https://maidsafe.net"
+license = "GPL-3.0"
+name = "parsec_dev_utils"
+repository = "https://github.com/maidsafe/parsec"
+version = "0.0.1"
+
+[dependencies]
+maidsafe_utilities = "~0.17.0"
+parsec = { path = ".." }
+proptest = "~0.8.6"
+rand = "~0.4.2"
+unwrap = "~1.2.0"
