+++ tests/integration_tests.rs
@@ -56,15 +56,10 @@
 
 #[macro_use]
 extern crate log;
-extern crate maidsafe_utilities;
-extern crate parsec;
+extern crate parsec_dev_utils as utils;
 extern crate rand;
 #[macro_use]
-extern crate unwrap;
-#[macro_use]
 extern crate proptest;
-
-mod utils;
 
 use self::utils::proptest::{arbitrary_delay, ScheduleOptionsStrategy, ScheduleStrategy};
 use self::utils::{
